requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Werkzeug==2.3.7
gunicorn==21.2.0
together==1.4.6
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5.1
markupsafe==2.1.3
jinja2==3.1.2